requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
nltk
spacy
python-docx
PyPDF2
sentence-transformers
gunicorn
pytest
pymongo